Fix review findings: cleanup paths, IMU capture determinism, pairing, LP64 bug, stale docs
All five findings verified against the code before fixing.

High - cam_direct could leave cameras powered and the FSIN strobe running on setup
failure. Both early `return 1` paths in main() sat after syncboss_cam_power(1) /
syncboss_lib_start(), and stage_capture() returned with cam A still started if cam B
failed to acquire. That second one compounded: main()'s tail only released g_sensor,
which stage_capture never sets, so qcamera_close would hit the "Cameras still in use!"
assert and abort, skipping syncboss teardown entirely. Now there is one cleanup path
(goto cleanup) plus a NULL-safe release_sensors() used on every exit from stage_capture.

High - IMU capture was not deterministic. The syncboss.raw open and pthread_create
returns were ignored, and the thread was never joined: it blocked in read(), so at exit
the stdio buffer for syncboss_chunks.csv was never flushed and the tail of syncboss.raw
could be lost -- precisely the data the next clock-map capture depends on. The thread now
uses poll() with a 100 ms timeout so it observes g_imu_run, both failures abort the
capture with cleanup, and the thread is joined before the fd is closed.

Medium - build_euroc_direct.py paired stereo frames and then snapped, deduped and
IMU-lead-filtered each camera independently. Equal lengths held for this capture by luck.
Snapping is now done once per PAIR and the paired structure is kept throughout, with
assertions that the two timestamp lists are identical and strictly increasing.

Medium - mempeek: %lu wrote 8 bytes into an `unsigned ino` on LP64 while parsing
/proc/<pid>/maps, which is UB and could corrupt adjacent stack locals. Now unsigned long.

Medium - stale docs: README said the first IMU frame was still pending (superseded --
IMU comes from the syncboss kernel FIFO at ~994 Hz); QWEN.md still framed syncboss-direct
as lacking the 1 kHz IMU and cameras (both now resolved via libsyncboss); and the runner
comment plus a notes/11 gotcha still said cfg[0]=42 when the resolved value is 112.

+++ tools/mempeek/mempeek.c
@@ -47,7 +47,8 @@ static void load_maps(const char *pid) {
     if (!f) die("open maps");
     char line[1024];
     while (fgets(line, sizeof line, f)) {
-        uint64_t s, e; char perms[8]; uint64_t off; unsigned long devmaj, devmin; unsigned ino;
+        uint64_t s, e; char perms[8]; uint64_t off; unsigned long devmaj, devmin; unsigned long ino;  // %lu writes 8 bytes on LP64 --
+                                                          // an `unsigned` here corrupted the stack
         char name[512] = {0};
         int n = sscanf(line, "%lx-%lx %4s %lx %lx:%lx %lu %511[^\n]",
                        &s, &e, perms, &off, &devmaj, &devmin, &ino, name);
